requirements.txt
django>=3.2
pillow
qrcode
openpyxl
pandas
python-dotenv
pymysql
